Fix ViewAnimatorTest's testConstructor Test
Remove a line in the test that expected the constructor to work with a null
Context. Also remove the subsequent line that tried to test two null
arguments, since throwing a NPE is not a behavior that needs to be expected
and enforced. Leave the remaining tests, since they do test that nothing
blows up when given all non-null arguments.

Change-Id: I0ae638310970fca061c9e6145f5a2b2e05d464e3