Catch RunTimeExceptions in SignatureTest
SignatureTest was throwing a RunTimeException, because the
current.xml has a private class. The exception caused
SignatureTestActivity to not return a result bundle, which
then caused CTS to claim the test timed out. Thus, catch
the exceptions and classify them under the existing
CAUGHT_EXCEPTION failure type. Add an extra parameter to
the load methods so that the error messages give some
sort've clue as to where the problem in the XML is.

Change-Id: I7006031420fc9901e11c0e5b400ab1ec2f0012e2