Fix KnownFailure TypefaceTest
This test was failing due to the generated image not matching the
expected image. It is very difficult to see any difference between
the two images though, because the generated one differs by 1 pixel
in certain areas. Replace the expected image with the current one
being generated and remove the KnownFailure annotation.

Change-Id: Iac4f14ef23f4f3edc5a78de6165ebfe1a3c9036a