Fix for b 2558942
Updated the test instead of fixing webview. I think there can be many reasons why some pixels are a bit off since this has been rendered to a real
display. This should check that the image is mostly the same, but allowing for some error.

Also improved the logic a bit.

Change-Id: I9d7f412f7109d4854f34556e38a3b4f3b460710b