View_AnimationTest QVGA Fixes
Bug 2597973

The animations time out, because the view being animated is not
visible on the screen. This is because the layout being used has
two views that are 400 pixels in height and push the animated
view down. Just use the first view in the layout.

Also incorporate a prior fix for testClearDuringAnimation which
was incorrectly asserting about the animation ending.

Change-Id: Icc5032131aa78a16ade0d73b0c83c2122d688ece