Add CtsTestCaseList.mk to Build CTS
Bug 2732160

Extract the test case list from build/core/tasks/cts.mk and put
them in a CtsTestCaseList.mk file. This change will make it
easier to develop CTS in the open, since adding test packages
and applications will no longer require multiple project changes.

Another alternative was to move the entire cts.mk file to the
CTS project but since the Makefile is in flux in the master branch
so much that I've decided to simplify things by only including
the test case list to avoid merge issues.

Also I decided to keep this in a separate Makefile rather than
stuffing the definitions in the CTS Android.mk. I'm thinking
its better to be more explicit with an include statement
rather than being implicitly defined in the Android.mk.

Change-Id: I7476ed02271d03fe1c6082e3e6f3013004d8ceae