Some PackageManager Feature Tests
Bug 2817090, 2817037, 2817010, 2817022, 2817050

Some sanity tests to check that the PackageManager is reporting
the correct features. It won't catch everything, and some of
the tests like for checking for the existence of WiFi or
Camera don't make a whole lot of sense... Finally, test that
features not defined by the SDK aren't using the android
namespace.

Moved feature tests for sensors and telephony into this class
to centralize all these feature tests. Thanks to George for
nearly all the code here.

Change-Id: I67e1e3ca8ccd8bfe46422f97abf571aecd28ad06