Read Log Permission Test Fixes
Bug 2964078

Fix a typo in the logcat command used in the test. There was an
extra colon in one of the arguments.

Add a test to check that log files under "/dev/log" are not world
readable. Also check they are owned by user root and group log.

Create a JNI library just for the permissions package. The JNI code
cannot be added to the CtsTestStubs JNI library, because the tests
cannot use CtsTestStubs since that acquires the permissions we
are trying to negative test. I also tried compiling just the single
Java source in CtsTestStubs (initially where I tried to put
FileUtils) and reusing the CtsTestStubs JNI library but the VM did
not like it. Thus, I take the easiest route of making a separate
JNI library just for the permisions package.

Change-Id: I6aafe9cfcdee1250f5d693cf83abb826faf63344