FocusFinderTest Workarounds...
Bug 2437360

The FocusFinder tests were failing on large width screens, because
the focus model takes into account the width of the widgets. The
prior layout had buttons that spanned the entire screen, and this
caused the focus behavior to change across different screen sizes.
This does not seem correct, but it is how the framework is coded.
Create a new layout that has four equally-sized center buttons to
avoid this problem.

Finally, the FocusFinder#findNearestTouchables seems to not take
into account the touch coordinates when calculating the delta
offset for the down and right focus directions which thus makes
the distance greater than the "edge slop" causing it to fail.
This also does not seem correct but probably due to a lack of
my understanding (Bug 2986136). Work around the problems using
more simple cases that touch on the outer edges of the widgets
that do not trigger that problem.

(Backport from Froyo. match_parent attributes changed to fill_parent
and strings added to strings.xml to get around translation errors)

Change-Id: I1b33e254c4d15c29f85713758ccf61366b6071da