CtsVerifier Test Info Dialogs
Add an "Info" button in between the "Pass" and "Fail" buttons
that shows a dialog with instructions about a test activity.

Show the info dialog when the user visits the test activity
for the first time. Use the existing content provider to
keep track of this.

Refactor the Suid File Scanner to be more user friendly by
showing an info dialog rather than start scanning immediately.
Show authorized SUID binaries with a green background and
unauthorized binaries with a red background. Fix a bug where
the dialog got leaked due to calling hide rather than dismiss.

Change-Id: I0626dc0634a592193f5e4d820c72a1541d52a8b0