Fix Eclair ARMEABI Test
The ro.product.cpu.abi2 property for non-ARMEABI_V7 devices was
incorrectly being checked for "unknown." The value of "unknown"
comes from the Froyo version of this test which is what the
Build.CPU_ABI2 constant was if the property was not set. Fix the
Eclair version by using getprop to make sure the property is
not set at all.

Change-Id: Ieaa22fbbc06b9349b4de0fcbb5fef55207c4728c