Improve Froyo ARMEABI Test
Unlike the Eclair port of the test, there wasn't anything
wrong with the Froyo version, because the Build.CPU_ABI2
returns Build.UNKNOWN when the property is not set. However,
the test couldn't make a distinction between the property
not being set and a property being set with the value "unknown."

Therefore, port the same enhancements done for the Eclair test
and improve the messaging.

Change-Id: Ie0edfc390065da0d6c75cc6490eb9d319ad8de91