CTS Test: Set fullscreenMode as false before checking.
[Comment]
The following test was failing during the view test cases.
android.view.inputmethod.cts.BaseInputConnectionTest#testReportFullscreenMode

Reason:
The mFullScreenMode is set as true when the test check for false status.

Solution:
This patch will set as false before it check for the false status as it
is implemented for true condition.

Change-Id: I7705cea0dd129564a85cc45aa210587076583897
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>