Fix Dates in MediaProvider Tests
The tests inserted a date added value and were checking that they
came back. However, these tests started failing, because the
MediaProvider itself now inserts a date added on its own. Thus,
the times were off also due to the fact that the provider uses
seconds rather than the test that was putting milliseconds.

The prior release also had code to insert the date added times.
However, it seems there was a bug where the wrong ContentValues
value was being passed to the final insert method which is why
it didn't work.

The tests were using milliseconds for the date modified, but the
documentation for the MediaStore keys say that seconds are used.
The tests were failing because the milliseconds were being read out
from the database as a truncated integer causing the failures.

Change-Id: I0ac982e3424a8103dd8deb36b1f7188e57b68219