Refactor Dialog Tests
Bug 3184944

The dialog tests start an activity that had a list of buttons
where each button launched a different type of dialog and then
hit the down button a number of times and then the center button
to launch them. Replace this approach with an intent based
approach where the tests launch an intent specifying what dialog
to create. This should make the tests less fragile to UI changes.

Change-Id: I3d36e2f9e6c1b4a065ffda59289a85278e04c8cf