Fix WebViewTest#testScrollBarOverlay
Bug 3227771

This test was occasionally timing out in waitForIdleSync, so you had
to change the device orientation to get it unstuck. Remove the flinging
checks that the test was doing and just stick to simple unit testing
of the setters and getters.

NOTE: Can't seem to reproduce the time out anymore, but still checking
      this in to stick to simpler unit testing.

Change-Id: Iea2648812cb0a0bd478bb82d5148e9e07093b2cf