Fix AutoCompleteTextViewTest
The dropdown doesn't show up if there aren't enough items in the list.
The "inflatePopup" method was setting the text to blank and then trying
to show the dropdown. As soon as that happened, mIsShowing was set to
true, but the handler for the popup dismisses it because it doesn't have
enough leters. Fix the tests to support this new behavior.

Change-Id: Icab9eafa7a8945522da4345a2617b6062fd7131b