Fix ScrollViewTest#testRequestChildRectangle...
The test sets up a scroll view with a layout with 3-5 widgets inside of
it. It picks a rectangle in the lower right of the scroll view that
shoul cause scrolling, because the vertical fading edge takes up some of
the height. However, the vertical fading edge isn't enabled by default
anymore, so no scrolling occurred. Enable scrolling in the test to make
it work again.

Change-Id: I8e1c64a971df5af0f8e71d8e3e1e4c97726be8d9