Fix InsetDrawableTest#testGetOpacity
Bug 2397630

BitmapDrawableTest#testMutate changed the constant state which affected
InsetDrawableTest#testGetOpacity. However, the test wasn't doing much of
anything, so I've changed it a bit to call setAlpha and check
getOpacity a couple times. This test shouldn't have a problem with
constant state, because it calls setAlpha before checking anything.

Change-Id: I1cca5cb18107a647e40c13a7c88c905856436595