Fix AccountManagerTest#testGetAuthTokenByFeatures
Bug 3412989

Change I5f2624 "Fix previous quick fix, LoginOptions is final." and
Change Ifb6a8b "Fix NPE, caller may not pass loginOptions" changed
whether or not the bundle passed was passed back to getAuthToken.

The test has four main tests:

The first test tries calls getAuthTokenByFeature using a null
activity. The AccountManager doesn't have any accounts (although
theres a call to addAccountExplicitly at the beginning of the
test), so it calls addAccount instead of getAuthToken (no idea
why). As a result, there is no auth token returned, so the
test passes.

The second test was failing because it expected the auth
token options returned to be null. The AccountManager has an
account this time around, so it calls getAuthToken instead.
AccountManager#getAuthToken passed null for the options to
AccountManagerServices#getAuthToken which has the new NPE
check cited above so that causes an options bundle to be
passed to the MockAuthenticator which records having
received a bundle! This causes the failure. Thus, the test
now checks that the options bundle returned is non-null.

The final test returns a null options. This is because now
the AccountManager has put the account in its cache. As a
result, it returns without calling the MockAuthenticator
and returns a custom bundle using data from the cache.
As a result, the test passes, because the MockAuthenticator
wasn't called.

Change-Id: I2d71fe30f6e1cabcf673bfb362dc9dc61a9cc1f9