Fix EventLogTest Flakiness
Issue 14896

The test was getting the current time and collecting events since
that time. However, the times in the event logs are usually recorded
faster, so events could get dropped. As a result, there were sleep
calls that needed to be increased and adjusted for various devices.

Fix these tests to gather events by looking for an event with a
certain payload instead. This should be more reliable, since the
payload is the current time from the test. This doesn't test the
event times as much though.

Change-Id: I4f6c9ad3ced39f2aa03c8c01b01b27f59948f8d0