Change max/min level range to allow the equalizer to be replaced
Audioeffect framework has a mechanism to replace the effect with
the enhanced version. Some equalizer can have a solid equalizing
implementation with appropriate band gains. Max/min level range
should not be limited with a fixed value. According to the OpenSL
ES 1.0.1 specification(http://www.khronos.org/opensles/), the
range returned by GetBandLevelRange must at least include 0mB.
So this fix changes to check if the maximum level range is not
negative and minimum level range is not positive.

Change-Id: I0145545bddc35db119c82817882ea26a35268645