Clean up WebView CTS test
waitForLoadcomplete() and assertLoadUrlSuccessfully() don't need to take
the WebView and timeout as parameters. The former is a member, the
latter is a constant.

Also, don't propagate the InterruptException that may be thrown by
Thread.Sleep(), as there's nothing we can do about this and it
clutters the API.

This will simplify later refactoring to call all WebView methods on
the UI thread.

Bug: 4340864
Change-Id: I7cd2a9011023345ac8be934e13b86ff1fb10c9d3