Test for DownloadManager Compliance
Adds a test to enforce CDD section 7.6.1 that says the
DownloadManager must be able to download individual
files of 55MB.

Modify the CtsTestServer to generate files of any size
in response to the download queries. The test downloads
a file of 55MB and checks that the file size is correct
both in the database and by the parcel file descriptor
returned from DownloadManager.

Change-Id: I82f17bfe5937fdd63a80fc71540537463509c241