CTS Verifier Sub Tests Implementation
Refactor the Bluetooth test to use the same test results storage
as the main test list. This will make it easy for future tests
to have subtests like the Bluetooth test one does. Also it will
be possible to send a report that has the results for the sub
tests since all the data is in one place.

Create subtests by adding a meta data with the tag "test_parent"
to specify the class that this test belongs to.

Change-Id: Id7ec5da311751bb8c1af7915137dacc89286a23c