Cannot test properly if a quality is 'unspecified'
If a test generates 'unspecified' as a password quality,
then it fails at a test case of password minimum length.
DevicePolicyManager won't save the length a setting file,
when the password quality is configured as 'unspecified'.
In that case, the test compares an actual min-length
with an empty expected value.
This test won't check any case exhaustively.
So I believe the 'unspecified' case can be omitted.

Change-Id: Id6a1284b77832691f6fd0761557f78811e272403