Improve a performance of a download file creation
CtsTestServer may take a long time on a slower device
when it creates 55MB download file. If the time was over 20 sec,
then Http request would be timed-out. In that case,
the test android.app.cts.DownloadManagerTest#testMinimumDownload fails.
(It's not 2min time-out in this test case.)

So it is needed to improve file write performance like this patch.

Change-Id: I1b153e02d15b4db786d407f5f2fa9b04566e9caa